Fix CEPH to be working and used by default
